gralloc: only use the private_handle_t structure that I imported in include/gralloc_priv.h for Exynos7870.
I had imported include/gralloc_priv.h from other older BSP sources, for use
on Exynos7870, and the only differences between the contents of that
gralloc_priv.h header and the contents of the other gralloc_priv.h header
that I replaced were in the private_handle_t structure. On Exynos7580 and
others, the private_handle_t structure from the gralloc_priv.h header that
I replaced needs to be used, rather than the private_handle_t structure from
the gralloc_priv.h header that I replaced it with.

Therefore, ifdef out the contents of the current private_handle_t structure
so that it's only used if the target SoC is exynos7870, and import the
private_handle_t structure that I had replaced alongside and after the
private_handle_t structure for Exynos7870 to be used if the target SoC
is otherwise not Exynos7870.

Change-Id: Ida9da0da91b03355759585135f30e5e9a138ba72